requirements.txt
dj-database-url==2.3.0
Django==5.1.5
djangorestframework==3.15.2
psycopg2==2.9.10
gunicorn==20.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2